init/set_up_project
