+++ package.json
@@ -9,9 +9,15 @@
     "@types/jest": "^27.5.2",
     "axios": "^1.2.2",
     "bootstrap": "^5.2.3",
+    "flowbite": "^1.6.1",
+    "flowbite-react": "^0.3.7",
+    "js-cookie": "^3.0.1",
     "node-sass": "^7.0.3",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
+    "react-icons": "^4.7.1",
+    "react-responsive-carousel": "^3.2.23",
+    "react-router-dom": "^6.6.2",
     "react-scripts": "5.0.1",
     "sass": "^1.57.1",
     "typescript": "^4.9.4",
