+++ src/components/Button/Button.tsx
@@ -1,0 +1,11 @@
+import React from 'react';
+
+const Button: React.FC = () => {
+    return (
+        <a href="/" className="btn btn-primary">
+            HAHA
+        </a>
+    );
+};
+
+export default Button;
