+++ src/helpers/image.ts
@@ -1,3 +1,3 @@
 export function checkURLIsImage(url: { match: (arg0: RegExp) => null }) {
-  return url?.match(/\.(jpeg|jpg|gif|png)$/) != null;
+    return url?.match(/\.(jpeg|jpg|gif|png)$/) != null;
 }
