+++ src/helpers/input.ts
@@ -1,0 +1,4 @@
+export const isValidInput = (value: string) => {
+    if (!value || value === '') return false;
+    return true;
+};
