--- src/helpers/input.tsx
@@ -1,4 +1,0 @@
-export const isValidInput = (value: string) => {
-  if (!value || value === "") return false;
-  return true;
-};
