+++ src/router/index.ts
@@ -1,0 +1,3 @@
+export const ROUTER = {
+    INDEX: '/',
+};
